requirements.txt
# ── Database ──────────────────────────────────────────
pyodbc==4.0.39
sqlalchemy==2.0.30

# ── Data processing ───────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Configuration ─────────────────────────────────────
python-dotenv==1.0.1

# ── Logging ───────────────────────────────────────────
loguru==0.7.2

# ── Dashboards (Phase 2) ──────────────────────────────
flask==3.0.3
jinja2==3.1.4

# ── PDF generation (Phase 3) ─────────────────────────
weasyprint==62.3

# ── Testing ───────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0
